refactor: split remaining backend core modules
