+++ backend/src/routes/media-core-admin-processing.ts
@@ -1,0 +1,10 @@
+import { Hono } from 'hono';
+import { registerMediaAdminProcessingUploadRoutes } from './media-core-admin-processing-upload';
+import { registerMediaAdminProcessingAIMRoutes } from './media-core-admin-processing-ai';
+
+const media = new Hono();
+
+registerMediaAdminProcessingUploadRoutes(media);
+registerMediaAdminProcessingAIMRoutes(media);
+
+export default media;
