+++ docs/dev-logs/2026-06-06-backend-core-final-pass.md
@@ -1,0 +1,17 @@
+# Backend Core Final Pass
+
+Date: 2026-06-06
+
+## Summary
+- Broke down the remaining large backend files into smaller route and utility modules.
+- Reduced the remaining large files to lower-risk boundaries.
+- Preserved behavior and re-verified the backend and frontend build/test matrix.
+
+## Highlights
+- Split `media-core-admin-processing` into upload/transcribe and summarize/extract modules.
+- Split `ai-controls-quotas` into schema and operations modules.
+- Split `smart-chat-core` and `ai-engine-utils` into barrel + focused implementation modules.
+
+## Verification
+- `npm run verify` passed
+- `npm run test:backend` passed
